Fix 64-bit build
* Update Paladin project file to 2.0.
* Last gcc2 compatible commit.

+++ src/general/BitmapView.cpp
@@ -656,7 +656,7 @@ BitmapView::BitmapFromClipboard(void)
 			be_clipboard->Unlock();
 			
 			const void *buffer;
-			int32 bufferLength;
+			ssize_t bufferLength;
 			
 			BRect frame;
 			color_space cspace = B_NO_COLOR_SPACE;
@@ -688,7 +688,7 @@ BitmapView::BitmapFromClipboard(void)
 		
 		case CLIP_BEOS: {
 			const void *buffer;
-			int32 bufferLength;
+			ssize_t bufferLength;
 			
 			BRect frame;
 			color_space cspace = B_NO_COLOR_SPACE;
